fixed
If the user's position is beyond the maximum number of boards, it is modified to move to the last column and output that it is finished at the end of the board. And if the number of survivors is 0, and the winner is -1, no one survives, so there is no winner. Finally, it can be checked by reflecting the player's status. Modified to not double-check players who are already dead